Dismantle `CommonFunction`
