+++ shared/src/main/scala/pt/kcry/blake3/CompressRounds.scala
@@ -11,9 +11,8 @@
 
 package pt.kcry.blake3
 
-private[blake3] object CommonFunction {
-  @inline
-  private def compressRounds(
+private[blake3] object CompressRounds {
+  def compressRounds(
     state: Array[Int], blockWords: Array[Int], chainingValue: Array[Int],
     counter: Long, blockLen: Int, flags: Int
   ): Unit = {
@@ -977,188 +976,4 @@ private[blake3] object CommonFunction {
     state(14) = state_14
     state(15) = state_15
   }
-
-  def compressInPlace(
-    chainingValue: Array[Int], bytes: Array[Byte], bytesOffset: Int,
-    counter: Long, blockLen: Int, flags: Int, tmpState: Array[Int],
-    tmpBlockWords: Array[Int]
-  ): Unit = {
-
-    tmpBlockWords(0) = ((bytes(3 + bytesOffset) & 0xff) << 24) |
-      ((bytes(2 + bytesOffset) & 0xff) << 16) |
-      ((bytes(1 + bytesOffset) & 0xff) << 8) | ((bytes(0 + bytesOffset) & 0xff))
-
-    tmpBlockWords(1) = ((bytes(7 + bytesOffset) & 0xff) << 24) |
-      ((bytes(6 + bytesOffset) & 0xff) << 16) |
-      ((bytes(5 + bytesOffset) & 0xff) << 8) | ((bytes(4 + bytesOffset) & 0xff))
-
-    tmpBlockWords(2) = ((bytes(11 + bytesOffset) & 0xff) << 24) |
-      ((bytes(10 + bytesOffset) & 0xff) << 16) |
-      ((bytes(9 + bytesOffset) & 0xff) << 8) | ((bytes(8 + bytesOffset) & 0xff))
-
-    tmpBlockWords(3) = ((bytes(15 + bytesOffset) & 0xff) << 24) |
-      ((bytes(14 + bytesOffset) & 0xff) << 16) |
-      ((bytes(13 + bytesOffset) & 0xff) << 8) |
-      ((bytes(12 + bytesOffset) & 0xff))
-
-    tmpBlockWords(4) = ((bytes(19 + bytesOffset) & 0xff) << 24) |
-      ((bytes(18 + bytesOffset) & 0xff) << 16) |
-      ((bytes(17 + bytesOffset) & 0xff) << 8) |
-      ((bytes(16 + bytesOffset) & 0xff))
-
-    tmpBlockWords(5) = ((bytes(23 + bytesOffset) & 0xff) << 24) |
-      ((bytes(22 + bytesOffset) & 0xff) << 16) |
-      ((bytes(21 + bytesOffset) & 0xff) << 8) |
-      ((bytes(20 + bytesOffset) & 0xff))
-
-    tmpBlockWords(6) = ((bytes(27 + bytesOffset) & 0xff) << 24) |
-      ((bytes(26 + bytesOffset) & 0xff) << 16) |
-      ((bytes(25 + bytesOffset) & 0xff) << 8) |
-      ((bytes(24 + bytesOffset) & 0xff))
-
-    tmpBlockWords(7) = ((bytes(31 + bytesOffset) & 0xff) << 24) |
-      ((bytes(30 + bytesOffset) & 0xff) << 16) |
-      ((bytes(29 + bytesOffset) & 0xff) << 8) |
-      ((bytes(28 + bytesOffset) & 0xff))
-
-    tmpBlockWords(8) = ((bytes(35 + bytesOffset) & 0xff) << 24) |
-      ((bytes(34 + bytesOffset) & 0xff) << 16) |
-      ((bytes(33 + bytesOffset) & 0xff) << 8) |
-      ((bytes(32 + bytesOffset) & 0xff))
-
-    tmpBlockWords(9) = ((bytes(39 + bytesOffset) & 0xff) << 24) |
-      ((bytes(38 + bytesOffset) & 0xff) << 16) |
-      ((bytes(37 + bytesOffset) & 0xff) << 8) |
-      ((bytes(36 + bytesOffset) & 0xff))
-
-    tmpBlockWords(10) = ((bytes(43 + bytesOffset) & 0xff) << 24) |
-      ((bytes(42 + bytesOffset) & 0xff) << 16) |
-      ((bytes(41 + bytesOffset) & 0xff) << 8) |
-      ((bytes(40 + bytesOffset) & 0xff))
-
-    tmpBlockWords(11) = ((bytes(47 + bytesOffset) & 0xff) << 24) |
-      ((bytes(46 + bytesOffset) & 0xff) << 16) |
-      ((bytes(45 + bytesOffset) & 0xff) << 8) |
-      ((bytes(44 + bytesOffset) & 0xff))
-
-    tmpBlockWords(12) = ((bytes(51 + bytesOffset) & 0xff) << 24) |
-      ((bytes(50 + bytesOffset) & 0xff) << 16) |
-      ((bytes(49 + bytesOffset) & 0xff) << 8) |
-      ((bytes(48 + bytesOffset) & 0xff))
-
-    tmpBlockWords(13) = ((bytes(55 + bytesOffset) & 0xff) << 24) |
-      ((bytes(54 + bytesOffset) & 0xff) << 16) |
-      ((bytes(53 + bytesOffset) & 0xff) << 8) |
-      ((bytes(52 + bytesOffset) & 0xff))
-
-    tmpBlockWords(14) = ((bytes(59 + bytesOffset) & 0xff) << 24) |
-      ((bytes(58 + bytesOffset) & 0xff) << 16) |
-      ((bytes(57 + bytesOffset) & 0xff) << 8) |
-      ((bytes(56 + bytesOffset) & 0xff))
-
-    tmpBlockWords(15) = ((bytes(63 + bytesOffset) & 0xff) << 24) |
-      ((bytes(62 + bytesOffset) & 0xff) << 16) |
-      ((bytes(61 + bytesOffset) & 0xff) << 8) |
-      ((bytes(60 + bytesOffset) & 0xff))
-
-    compressRounds(tmpState, tmpBlockWords, chainingValue, counter, blockLen,
-      flags)
-
-    chainingValue(0) = tmpState(0) ^ tmpState(8)
-    chainingValue(1) = tmpState(1) ^ tmpState(9)
-    chainingValue(2) = tmpState(2) ^ tmpState(10)
-    chainingValue(3) = tmpState(3) ^ tmpState(11)
-    chainingValue(4) = tmpState(4) ^ tmpState(12)
-    chainingValue(5) = tmpState(5) ^ tmpState(13)
-    chainingValue(6) = tmpState(6) ^ tmpState(14)
-    chainingValue(7) = tmpState(7) ^ tmpState(15)
-
-  }
-
-  def compressInPlace(
-    state: Array[Int], chainingValue: Array[Int], blockWords: Array[Int],
-    counter: Long, blockLen: Int, flags: Int
-  ): Unit = {
-    compressRounds(state, blockWords, chainingValue, counter, blockLen, flags)
-
-    state(0) ^= state(8)
-    state(1) ^= state(9)
-    state(2) ^= state(10)
-    state(3) ^= state(11)
-    state(4) ^= state(12)
-    state(5) ^= state(13)
-    state(6) ^= state(14)
-    state(7) ^= state(15)
-
-    state(8) ^= chainingValue(0)
-    state(9) ^= chainingValue(1)
-    state(10) ^= chainingValue(2)
-    state(11) ^= chainingValue(3)
-    state(12) ^= chainingValue(4)
-    state(13) ^= chainingValue(5)
-    state(14) ^= chainingValue(6)
-    state(15) ^= chainingValue(7)
-  }
-
-  def compressSingle(
-    chainingValue: Array[Int], blockWords: Array[Int], counter: Long,
-    blockLen: Int, flags: Int
-  ): Int = {
-    val state = new Array[Int](BLOCK_LEN_WORDS)
-
-    compressRounds(state, blockWords, chainingValue, counter, blockLen, flags)
-
-    // a fast-track for single int
-    state(0) ^ state(8)
-  }
-
-  def compressSingleLong(
-    chainingValue: Array[Int], blockWords: Array[Int], counter: Long,
-    blockLen: Int, flags: Int
-  ): Long = {
-    val state = new Array[Int](BLOCK_LEN_WORDS)
-
-    compressRounds(state, blockWords, chainingValue, counter, blockLen, flags)
-
-    // a fast-track for single long
-    ((state(0) ^ state(8)).toLong << 32) | (state(1) ^ state(9))
-  }
-  def wordsFromLittleEndianBytes(bytes: Array[Byte]): Array[Int] = {
-    val res = new Array[Int](bytes.length / 4)
-    var i = 0
-    var off = 0
-    while (i < res.length) {
-      res(i) = ((bytes(3 + off) & 0xff) << 24) |
-        ((bytes(2 + off) & 0xff) << 16) | ((bytes(1 + off) & 0xff) << 8) |
-        ((bytes(off) & 0xff))
-      i += 1
-      off += 4
-    }
-    res
-  }
-
-  @inline
-  private def mergeChildCV(
-    merged: Array[Int], leftChildCV: Array[Int], rightChildCv: Array[Int]
-  ): Unit = {
-    System.arraycopy(rightChildCv, 0, merged, KEY_LEN_WORDS, KEY_LEN_WORDS)
-    System.arraycopy(leftChildCV, 0, merged, 0, KEY_LEN_WORDS)
-  }
-
-  def parentOutput(
-    blockWords: Array[Int], leftChildCV: Array[Int], rightChildCv: Array[Int],
-    key: Array[Int], flags: Int
-  ): Output = {
-    mergeChildCV(blockWords, leftChildCV, rightChildCv)
-    new Output(key, blockWords, 0, BLOCK_LEN, flags | PARENT)
-  }
-
-  def parentCV(
-    parentCV: Array[Int], leftChildCV: Array[Int], rightChildCv: Array[Int],
-    key: Array[Int], flags: Int, tmpBlockWords: Array[Int]
-  ): Unit = {
-    mergeChildCV(tmpBlockWords, leftChildCV, rightChildCv)
-    compressInPlace(parentCV, key, tmpBlockWords, 0, BLOCK_LEN, flags | PARENT)
-  }
-
 }
